+++ packages/jaspr/lib/src/components/html/html.dart
@@ -9,6 +9,7 @@ part 'forms.dart';
 part 'media.dart';
 part 'other.dart';
 part 'svg.dart';
+part 'table.dart';
 part 'text.dart';
 
 final _events = events;